Exchange chain info in networking handshake (#92)
* Exchange chain info in networking handshake

This will now send new fields along in the networking handshake and will get stored on the peer on the other side.
 * Heaviest head hash
 * Heaviest head sequence
 * Heaviest head cumulative work

Version has also changed to just an integer, and agent is a new handshake message field that is now freeform entry, and not required to be in a specific format by their peers. Instead of forcing exact versioning, we now enforce that the peer version is at least `VERSION_PROTOCOL_MIN`

These also now get send in the getPeers() endpoint. They will be used during the new block syncer algorithm.

* Update ironfish/src/network/peers/peerManager.ts

Co-authored-by: Derek Guenther <derek@ironfish.network>

* Add missing fields to verification

Co-authored-by: Derek Guenther <derek@ironfish.network>